Teamworking Login Page
This contains all my login page scripts.
It includes the HTML webpage and the php script for reading the credentials.
it also includes the css styling and images.
IMPORTANT: You will need to edit the "sql.php" script to user YOUR database credentials.
IMPORTANT: The "homepage.php" file included is just for testing purposes. It should be replaced with Aiden's homepage php script when it's created.